fix: 재고 데이터 소수점 보존 — Long→Double 타입 변경 + 프론트엔드 소수점 표시
Task 33: SAP RFC_002에서 소수점 2자리 재고 데이터(예: 2,106.00)를 보내지만
Long 타입으로 저장하면서 소수점이 잘리는 문제

수정:
1. PlantStorageLocation 엔티티
   - beginningInventory, availableInventory, availableStock, currentStock: Long → Double

2. SnopRecord 엔티티
   - beginningInventory, availableInventory: Long → Double

3. PlantStorageLocationDto, SnopRecordDto
   - 동일 필드 Long → Double

4. RfcReceiverService.processRfc002()
   - getLong → getDouble로 변경 (beginning_inventory, available_inventory)
   - 합산 Map<String, Long> → Map<String, Double>로 변경

5. 프론트엔드 formatNumber()
   - 소수점이 있는 값은 소수점 2자리까지 표시
   - 정수는 기존과 동일하게 소수점 없이 표시

- cache bust: v=20260529e

+++ module-sales/src/main/java/com/company/module/sales/dto/PlantStorageLocationDto.java
@@ -34,9 +34,9 @@ public class PlantStorageLocationDto {
     @Size(max = 10)
     private String stockUnit;
 
-    private Long beginningInventory;
-    private Long availableInventory;
-    private Long availableStock;
-    private Long currentStock;
+    private Double beginningInventory;
+    private Double availableInventory;
+    private Double availableStock;
+    private Double currentStock;
     private LocalDateTime sapSyncAt;
 }
